ai-skill-up-india/requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
google-generativeai
